x86_64: fix POSIX test-suite failures (test53/64/66/74/76) + ffs tweaks
Resolve several minix-posix failures on amd64:

- test53: u64.h ex64lo/ex64hi returned unsigned long (64-bit on amd64) and
  no longer truncated; mask through (u32_t) while keeping the unsigned long
  return type (preserves i386 semantics and %lx callers).
- test66: use int32_t/uint32_t for the 32-bit math locals; unsigned long is
  64-bit on LP64 and changed (unsigned long)(-10) % 7.
- test74: kernel-mediated copies must honor read-only/COW page protection.
  (1) createpde() takes a 'writable' flag and returns the not-present
  sentinel 0 when a write targets a leaf lacking AMD64_VM_WRITE, so it routes
  through vm_suspend(writeflag=1) for COW or EFAULT; (2) call vm_enable_paging()
  on the amd64 UEFI path so CR0.WP is set (it was skipped); (3) vm_memset()
  (safememset, used by /dev/zero) got the same explicit not-present handling.
- test64: add a VM low-memory reserve for system services (PAGES_RESERVED,
  free_pages_cnt, PAF_USERMEM via VR_USERMEM/vm_isuserp) so a user process
  exhausting memory can no longer starve VFS into a fatal fault and panic;
  user mmap now returns MAP_FAILED cleanly before exhaustion.
- test76: silence an upstream debug printf in vfs/worker.c.

Also: register the ffs service in etc/system.conf, guard fs_sync() against a
NULL superblock during a failed-mount teardown, and improve read_super()'s
UFS1-not-supported message.

Docs: kernel-arch-x86_64.md (RO/COW kernel copies, test77 open issue) and
vm-x86_64-port.md (low-memory reserve).

test77 (PTY) remains an open pre-existing stack/COW heisenbug; documented.

+++ minix/tests/test66.c
@@ -10,6 +10,7 @@
  */
 
 #include <stdio.h>
+#include <stdint.h>
 #include <assert.h>
 
 int max_error = 10;
@@ -41,8 +42,8 @@ int main(int argc, char *argv[])
 			double da = a/4.0, db = b/4.0;
 			signed long long a64s = a, b64s = b, ds, ms;
 			unsigned long long a64u = a, b64u = b, mu;
-			signed long a32s = a, b32s = b, ds32, ms32;
-			unsigned long a32u = a, b32u = b, mu32;
+			int32_t a32s = a, b32s = b, ds32, ms32;
+			uint32_t a32u = a, b32u = b, mu32;
 
 			/* indicate no result */
 			mu32 = ds32 = ms32 = mu = ds = ms = 31337;
